Update MIPS Assignment 1
This commit contains a temporary section of code used to test the operation of reading in an input string. The code displays the string that was entered as output in order to check that the string was correctly read as an input. This section of code will be removed at a later time in order to follow the description of the assignment.