Update MIPS Assignment 1
This commit contains the loop and branch statements that loop through each character in the input string and then checks the character to see if it is between '0' and '9', 'a' and 'f', and 'A' and 'F'. If all of the characters in the input string are between these specified characters, then the code continues to the next section of the assignment. Otherwise, it outputs the error message, "Invalid Hexadecimal Number".